requirements.txt
requests
python-dotenv
datetime
groq